Ensure we dont get stuck in an infinite loop of sending SIGTERM to ourselves
git-svn-id: http://svn.open-mpi.org/svn/orcm/trunk@179 e75c48be-c379-0410-a98d-d1e110c868e2

+++ src/orte/mca/errmgr/orcmapp/errmgr_orcmapp.c
@@ -31,6 +31,7 @@
 #include "orte/util/name_fns.h"
 #include "orte/runtime/orte_globals.h"
 #include "orte/mca/errmgr/errmgr.h"
+#include "orte/mca/ess/ess.h"
 
 #include "mca/pnp/pnp.h"
 #include "mca/leader/leader.h"
@@ -89,6 +90,7 @@ orte_errmgr_base_module_t orte_errmgr_orcmapp_module = {
  * Local functions and globals
  */
 static orte_thread_ctl_t ctl;
+static bool kill_sent=false;
 
 static void notify_failure(int status,
                           orte_process_name_t *sender,
@@ -152,7 +154,16 @@ static void app_abort(int error_code, char *fmt, ...)
     }
     va_end(arglist);
     
-    kill(getpid(), SIGTERM);
+    if (kill_sent) {
+        /* only send SIGTERM to ourselves once as
+         * we otherwise can get into an infinite loop
+         * while trying to abnormally terminate
+         */
+        orte_ess.abort(error_code, false);
+    } else {
+        kill_sent = true;
+        kill(getpid(), SIGTERM);
+    }
     return;
 }
