+++ src/orte/mca/errmgr/orcmsched/errmgr_sched.c
@@ -71,6 +71,7 @@
 #include "orte/mca/routed/routed.h"
 #include "orte/mca/sensor/sensor.h"
 #include "orte/orted/orted.h"
+#include "orte/mca/ess/ess.h"
 
 #include "mca/pnp/pnp.h"
 #include "util/triplets.h"
@@ -126,6 +127,7 @@ orte_errmgr_base_module_t orte_errmgr_orcmsched_module = {
  */
 static orte_thread_ctl_t ctl;
 static orte_job_t *daemon_job=NULL;
+static bool kill_sent=false;
 static void recover_procs(orte_process_name_t *daemon_that_failed);
 static void remote_update(int status,
                           orte_process_name_t *sender,
@@ -194,7 +196,16 @@ static void sched_abort(int error_code, char *fmt, ...)
     }
     va_end(arglist);
     
-    kill(getpid(), SIGTERM);
+    if (kill_sent) {
+        /* only send SIGTERM to ourselves once as
+         * we otherwise can get into an infinite loop
+         * while trying to abnormally terminate
+         */
+        orte_ess.abort(error_code, false);
+    } else {
+        kill_sent = true;
+        kill(getpid(), SIGTERM);
+    }
     return;
 }
